Causal Distillation Trees

TODO

Usage

causalDT(
  X,
  Y,
  Z,
  holdout_idxs = NULL,
  holdout_prop = 0.3,
  teacher_model = "causal_forest",
  teacher_predict = NULL,
  student_model = "rpart",
  rpart_control = NULL,
  nfolds_crossfit = NULL,
  nreps_crossfit = NULL,
  B_stability = 100,
  max_depth_stability = NULL,
  ...
)

Arguments

X

A tibble, data.frame, or matrix of covariates.

Y

A vector of outcomes.

Z

A vector of treatments.

holdout_idxs

A vector of indices to hold out for honest estimate of treatment effects. If NULL, a holdout set of size holdout_prop x nrow(X) is randomly selected.

holdout_prop

Proportion of data to hold out for honest estimate of treatment effects. Used only if holdout_idxs is NULL.

teacher_model

A function that takes in the named arguments X, Y, Z, and (optional) additional arguments and returns a model object that can be used to predict individual-level treatment effects using teacher_predict(teacher_model, data).

rpart_control

A list of control parameters for the rpart function. See ? rpart.control for details.

nfolds_crossfit

Number of folds for crossfitting.